+++ runner/src/server/plugins/engine/components/CheckboxesField.ts
@@ -1,10 +1,10 @@
 import { FormData, FormSubmissionErrors, FormSubmissionState } from "../types";
-import { ListFormComponent } from "./ListFormComponent";
 import { ListComponentsDef } from "@xgovformbuilder/model";
 import { FormModel } from "../models";
 import joi from "joi";
+import { SelectionControlField } from "server/plugins/engine/components/SelectionControlField";
 
-export class CheckboxesField extends ListFormComponent {
+export class CheckboxesField extends SelectionControlField {
   constructor(def: ListComponentsDef, model: FormModel) {
     super(def, model);
 
@@ -22,53 +22,17 @@ export class CheckboxesField extends ListFormComponent {
     this.stateSchema = schema;
   }
 
-  getAnswers(state: FormSubmissionState) {
-    const { name } = this;
-    const value = state[name];
-    const checked = Array.isArray(value) ? value : [value];
-    return checked
-      .map((check) => {
-        const item = this.items.find((item) => item.value === check);
-        return item?.value;
-      })
-      .filter((c) => c);
-  }
-
   getDisplayStringFromState(state: FormSubmissionState) {
     return state?.[this.name]?.join(", ");
   }
 
   getViewModel(formData: FormData, errors: FormSubmissionErrors) {
-    const { name } = this;
-    let viewModel = super.getViewModel(formData, errors);
-    let answers = formData[name] ?? [];
-    let formDataItems = Array.isArray(formData) ? answers : [answers];
-
-    viewModel.fieldset = {
-      legend: viewModel.label,
-    };
-
-    viewModel.items = viewModel.items.map((item) => {
-      const itemModel = {
-        ...item,
-        checked: !!formDataItems.find((i) => `${item.value}` === i),
-      };
-      if ("bold" in this.options && this.options.bold) {
-        itemModel.label = {
-          classes: "govuk-label--s",
-        };
-      }
-      if (item.hint) {
-        itemModel.hint = {
-          html: this.localisedString(item.hint),
-        };
-      }
-      /**
-       * TODO:- reintroduce addConditionalComponents when conditional reveals are fixed by GDS.
-       * //return super.addConditionalComponents(item, itemModel, formData, errors);
-       */
-      return itemModel;
-    });
+    const viewModel = super.getViewModel(formData, errors);
+    let formDataItems = (formData[this.name] ?? "").split(",");
+    viewModel.items = (viewModel.items ?? []).map((item) => ({
+      ...item,
+      checked: !!formDataItems.find((i) => `${item.value}` === i),
+    }));
 
     return viewModel;
   }
